[프로그래머스] 부족한 금액 계산하기

+++ Programmers/위클리 챌린지/1주차/부족한 금액 계산하기/Sanghoo.java
@@ -7,7 +7,7 @@ public class Sanghoo {
 
     public long solution(int price, int money, int count) {
         long answer = -1;
-        int totalPrice = 0;
+        long totalPrice = 0;
 
         for(int i=1; i<=count; i++) {
             totalPrice += price * i;
